Split out bits of upgrade.rs into upgrade_state.rs
